fix: update logic imports to account for folder rename
