+++ frontend/src/features/admin-form/create/common/components/CreatePageContent.tsx
@@ -3,12 +3,17 @@ import {
   DrawerTabs,
   useCreatePageDrawer,
 } from '~features/admin-form/create/CreatePageDrawerContext'
-import CreatePageLogicTab from '~features/admin-form/create/logic'
+
+import CreatePageLogicTab, { BuilderLogicProvider } from '../../logic'
 
 export const CreatePageContent = (): JSX.Element => {
   const { activeTab } = useCreatePageDrawer()
   if (activeTab === DrawerTabs.Logic) {
-    return <CreatePageLogicTab />
+    return (
+      <BuilderLogicProvider>
+        <CreatePageLogicTab />
+      </BuilderLogicProvider>
+    )
   }
 
   return <BuilderAndDesign />
