+++ frontend/src/features/admin-form/create/logic/index.ts
@@ -1,1 +1,2 @@
+export { BuilderLogicProvider } from './BuilderLogicContext'
 export { CreatePageLogicTab as default } from './CreatePageLogicTab'
